Addressed a bug in the rust fabrics demo app. Also took the opportunity to clean-up duplicate add/remove target and make progress function by sharing them for both Grain and Samples state.
